+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Install Banner PWA.
+
 ## [1.4.2] - 2019-08-07
 
 ### Fixed